+++ MTO/src/app/app.component.html
@@ -52,7 +52,6 @@
   </div>
 </div>
 
-
 <div class="content">
   <div class="content-pathAndButton">
     <div class="content-path">
@@ -73,6 +72,5 @@
       <img src="assets/img/gearWhite.svg">
     </button>
   </div>
-
-  <app-categories></app-categories>
+  <router-outlet></router-outlet>
 </div>
