--- MTO/src/app/categories/categories.component.html
@@ -1,93 +1,0 @@
-<div class="content-title">[Категории ОС]</div>
-
-  <div class="addButton-container">
-    <button class="addButton">
-      <span>Добавить</span>
-      <img src="assets/img/plus.svg">
-    </button>
-  </div>
-
-  <div class="table">
-    <div class="table-cells">
-      <div class="table-row table-header">
-        <div>[Категория ОС]</div>
-        <div>[Порядковый номер отображения категорий ОС]</div>
-        <div></div>
-      </div>
-      <div class="table-row">
-        <div>[Персональный компьютер]</div>
-        <div>[1]</div>
-        <div>
-          <img src="assets/img/bin.svg">
-        </div>
-      </div>
-      <div class="table-row">
-        <div>[Монитор]</div>
-        <div>[2]</div>
-        <div>
-          <img src="assets/img/bin.svg">
-        </div>
-      </div>
-      <div class="table-row">
-        <div>[Доска]</div>
-        <div>[3]</div>
-        <div>
-          <img src="assets/img/bin.svg">
-        </div>
-      </div>
-      <div class="table-row">
-        <div>[Стол]</div>
-        <div>[4]</div>
-        <div>
-          <img src="assets/img/bin.svg">
-        </div>
-      </div>
-    </div>
-    <div class="table-paginator">
-      <div class="paginator-buttons">
-        <div class="paginator-button">
-          <img src="assets/img/arrow-end-left.svg">
-        </div>
-        <div class="paginator-button">
-          <img src="assets/img/arrow-60-left.svg">
-        </div>
-        <div class="paginator-button current">
-          <span>1</span>
-        </div>
-        <div class="paginator-button">
-          <span>2</span>
-        </div>
-        <div class="paginator-button">
-          <span>3</span>
-        </div>
-        <div class="paginator-button">
-          <span>4</span>
-        </div>
-        <div class="paginator-button">
-          <span>5</span>
-        </div>
-        <div class="paginator-button">
-          <img src="assets/img/arrow-60-right.svg">
-        </div>
-        <div class="paginator-button">
-          <img src="assets/img/arrow-end-right.svg">
-        </div>
-      </div>
-
-      <div class="paginator-info">
-        <span>[1]</span>
-        <span> - </span>
-        <span>[4]</span>
-        <span> из </span>
-        <span>[4]</span>
-        <span> элементов</span>
-      </div>
-    </div>
-  </div>
-
-  <div class="addButton-container">
-    <button class="addButton">
-      <span>Добавить</span>
-      <img src="assets/img/plus.svg">
-    </button>
-  </div>
