+++ MTO/src/app/handbook_pages/categories/categories.component.scss
@@ -1,0 +1,147 @@
+// CONTENT TITLE
+.content-title {
+    color: #525252;
+    font-family: Montserrat, Roboto, sans-serif;
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: normal;
+    margin-top: 10px;
+    margin-bottom: 24px;
+}
+
+// CONTENT ADD BUTTON
+.addButton-container {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    padding: 12px;
+    border-bottom: 1px solid #C5C5C5;
+    background: #FFF;
+}
+
+.addButton {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 18px;
+    border: 0;
+    border-radius: 4px;
+    background: #9DD878;
+    padding: 7px 12px;
+    color: #FFF;
+    text-align: center;
+    font-family: Roboto, sans-serif;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 20px;
+    letter-spacing: 1.25px;
+    text-transform: uppercase;
+}
+
+// Это кнопка крч. Для теста сделал
+.addButton:active {
+    background: #1976d2;
+}
+
+// Конец теста
+
+
+// TABLE
+.table-row {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    border-bottom: 1px solid #C5C5C5;
+    background: #FFF;
+}
+
+.table-row div {
+    color: rgba(0, 0, 0, 0.87);
+    text-align: center;
+    font-family: Montserrat, Roboto, sans-serif;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 28px;
+    display: flex;
+    justify-content: center;
+    flex-grow: 1;
+    flex-basis: 50%;
+    padding: 10px 24px;
+    box-sizing: border-box;
+}
+
+.table-row div:last-child {
+    flex-grow: 0;
+    width: 80px;
+    flex-basis: unset;
+    flex-shrink: 0;
+}
+
+.table-header div {
+    color: rgba(0, 0, 0, 0.54);
+    line-height: 24px;
+    padding-top: 16px;
+    padding-bottom: 16px;
+}
+
+// PAGINATOR
+.table-paginator {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    padding: 10px 0;
+    border-bottom: 1px solid #C5C5C5;
+    background: #FFF;
+}
+
+.paginator-buttons {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.paginator-button {
+    display: flex;
+    width: 38px;
+    height: 38px;
+    justify-content: center;
+    align-items: center;
+    border-radius: 50%;
+    color: #C0D1DA;
+    text-align: center;
+    font-family: Roboto, sans-serif;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 20px;
+    letter-spacing: 1.25px;
+    text-transform: uppercase;
+}
+
+.paginator-button.current {
+    background: #C0D1DA;
+    color: #FFF;
+}
+
+// Это кнопка крч. Для теста сделал
+.paginator-button:active {
+    background: #1976d2;
+}
+
+// Конец теста
+
+.paginator-info {
+    color: rgba(0, 0, 0, 0.54);
+    text-align: right;
+    font-family: Montserrat, Roboto, sans-serif;
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 20px;
+    letter-spacing: 0.25px;
+    white-space: nowrap;
+}
