+++ MTO/src/app/handbook_pages/categories/categories.component.spec.ts
@@ -1,0 +1,21 @@
+import {ComponentFixture, TestBed} from '@angular/core/testing';
+
+import {CategoriesComponent} from './categories.component';
+
+describe('CategoriesComponent', () => {
+    let component: CategoriesComponent;
+    let fixture: ComponentFixture<CategoriesComponent>;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [CategoriesComponent]
+        });
+        fixture = TestBed.createComponent(CategoriesComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});
