+++ MTO/src/app/handbook_pages/categories/categories.component.ts
@@ -1,0 +1,10 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'app-categories',
+    templateUrl: './categories.component.html',
+    styleUrls: ['./categories.component.scss']
+})
+export class CategoriesComponent {
+
+}
