+++ MTO/src/app/handbook_pages/os-usage/os-usage.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'os-usage',
+  templateUrl: './os-usage.component.html',
+  styleUrls: ['./os-usage.component.scss']
+})
+export class OsUsageComponent {
+
+}
